Update build process and support IE 10/11
This updates all rollup, babel and 3rd party dependencies in order to
provide support for IE10/11 that includes polyfills by core-js.

It also modifies the UMD and ESM/CJS builds to optimize for their
corresponding envs.

Users of GgEzVp should not be impacted by these changes, as most of them
are directed to the build process, but there were some codebase changes
related to some of the dependencies new versions (like nano-events not
using a constructor anymore).